+++ analyzers/src/SonarAnalyzer.VisualBasic/Trackers/VisualBasicPropertyAccessTracker.cs
@@ -22,6 +22,7 @@ using System.Linq;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.VisualBasic;
 using Microsoft.CodeAnalysis.VisualBasic.Syntax;
+using SonarAnalyzer.Extensions;
 
 namespace SonarAnalyzer.Helpers.Trackers
 {
